# openapi-spec.yaml
openapi: 3.1.0
info:
  title: API Gateway
  description: |
    Single external entry point for the Innowise microservice platform.

    **Authentication:** all endpoints except `/api/v1/register`, `/api/v1/auth/login`, and `/api/v1/auth/refresh`
    require a valid JWT access token in the `Authorization: Bearer <token>` header.

    **Internal endpoints** (`POST /api/v1/auth/credentials`, `POST /api/v1/auth/validate`) are not exposed
    through the Gateway and are therefore not listed here. They are reachable only on the
    internal Docker network.
  version: 1.0.0
servers:
  - url: http://localhost:8090
    description: Local development
  - url: http://api-gateway:8080
    description: Docker Compose (internal)
...
paths:
...
  /api/v1/register:
    post:
      tags: [Registration]
      summary: Register a new user
      description: |
        Gateway orchestrates a two-step registration:
        1. Creates a user record in User Service (`POST /api/v1/users`).
        2. Saves credentials in Auth Service (`POST /api/v1/auth/credentials`).

        If step 2 fails, Gateway compensates with a two-step rollback:
        first deactivates the record (`PATCH /api/v1/users/{userId}/deactivate`),
        then hard-deletes it (`DELETE /api/v1/users/{userId}`).
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Validation error (missing fields, password too short)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered (propagated from Auth Service; no compensation triggered)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Upstream service returned 5xx — details sanitized, not forwarded to caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Upstream service unreachable (connection refused or timeout)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─────────────────────────────────────────────────────────────
  # AUTHENTICATION — proxied to Auth Service
  # ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token issued by Auth Service. Lifetime 15 minutes.
  schemas:
    RegisterRequest:
      type: object
      required: [name, surname, birthDate, email, password]
      properties:
        name:
          type: string
          example: Pavel
        surname:
          type: string
          example: Nikiforov
        birthDate:
          type: string
          format: date
          example: '1995-06-15'
        email:
          type: string
          format: email
          example: pavel@example.com
        password:
          type: string
          minLength: 8
          example: secret123
    RegisterResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          example: 42
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
        refreshToken:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2026-06-13T10:00:00Z'
        status:
          type: integer
          example: 401
        error:
          type: string
          example: Unauthorized
        message:
          type: string
          example: Invalid or expired token
        path:
          type: string
          example: /api/v1/users/me